dev: move error types to another file + simply dump and load implementation
